be more cautious in archive uri generation and downloaded archive name
as of checker-241, the archive is hosted on files.me.com - this broke the fragile archive uri generation which assumed the on-page uri was a relative link, it now tests and acts accordingly. the files.me.com hosted file also makes use of the "Content-Disposition:" http header to set the filename to download as, by default curl doesn't use this so a file named 'mt0ve6' was being downloaded and the unarchiving process would get tripped up making assumptions on the naming. the Content-Disposition header is fetched and if an explicit filename is provided that is used, otherwise the filename is determined from the archive uri.